Fix grammar in settings page 1
In English, when you use a length of time as an adjective to describe a noun, you will use a hyphen and the singular form of the time unit (ref: https://data.grammarbook.com/blog/numbers/numbers-as-adjectives/). For example: "I'm going to take a 5-minute break."

This is as opposed to the adverbial expression to describe the same thing which doesn't use a hyphen and uses the plural form of the time unit. For example: "I'm going to take a break for 5 minutes."

Looking at this setting in the UI, I feel like it would look weird to have the hyphen because it would look like a minus sign right beside the plus sign. So I didn't make that change. However, when you read this setting out loud, it sounds weird in English to have the plural form of the units.